+++ trunk/IMesh/IMesh.Triangulate/TriangulationHandler.h
@@ -6,10 +6,10 @@
 class TriangulationHandler{
 	grid *modelgrid;
 	all_point_list *apl;
-	float ball_radius;
+	double ball_radius;
 
 public:
 	TriangulationHandler(){};
-	void Initialize(vector<norm_point> &nps, int isize, int jsize, int ksize, float gridwidth,float ballradius);
+	void Initialize(vector<norm_point> &nps, int isize, int jsize, int ksize, double gridwidth,double ballradius);
 	void triangulateModel();
 };
